Simplify copying and pasting commands from CONTRIBUTING.md into cmd.exe; etc.
A)

This patch makes it easier for users to copy and paste your suggested commands into a console window.

(Yes, the trailing carriage returns are important.  For a set of four commands to run automatically, the user must include a trailing carriage return on the clipboard before clicking "Paste".)

B)

The patch also makes the main text of the file slightly easier to read and understand.